Make git status an async command
Use PowerShell runspaces to move the `git status` calls for posh-git off
the main console thread. Basic structure is:

1. Create / setup a runspace
2. Keep a position of where the cursor currently is in prompt
3. Call the posh-git status methods in the runspace
4. Build up a console buffer of the status
5. Write to the console buffer with the status (even if the cursor is no
longer there)
6. Dispose of completed runspaces

Note that this works because git status is the last part of my prompt.
Othwerise, either reserve space for status, or make the entire prompt
async so it can be written in order.